+++ apps/creator-app/src/components/Assistant/AssistantLayout/AssistantNavigation/AssistantNavigationTokenUsage.component.tsx
@@ -1,3 +1,4 @@
+import { PlanType } from '@voiceflow/internal';
 import { QuotaNames } from '@voiceflow/realtime-sdk';
 import { tid } from '@voiceflow/style';
 import { BaseProps, PrimaryNavigation, Tooltip } from '@voiceflow/ui-next';
@@ -24,6 +25,7 @@ export const AssistantNavigationTokenUsage: React.FC<BaseProps> = ({ testID }) =
   const plan = useSelector(Workspace.active.planSelector);
   const isOnProTrial = useSelector(Workspace.active.isOnProTrialSelector);
   const isProWorkspace = useSelector(Workspace.active.isProSelector);
+  const isTeamsWorkspace = plan === PlanType.TEAM;
 
   const refreshWorkspaceQuotaDetails = useDispatch(Workspace.refreshWorkspaceQuotaDetails);
 
@@ -73,11 +75,18 @@ export const AssistantNavigationTokenUsage: React.FC<BaseProps> = ({ testID }) =
         // show upgrade to pro button if user is on pro trial or is on starter plan
         .when(
           () => isOnProTrial || isStarterPlan(plan),
-          () => <Tooltip.Button size="small" label="Upgrade to pro" testID={tid(testID, 'upgrade')} onClick={() => paymentModal.openVoid({})} />
+          () => (
+            <Tooltip.Button
+              size="small"
+              label="Upgrade to pro"
+              testID={tid(testID, 'upgrade')}
+              onClick={() => paymentModal.openVoid({})}
+            />
+          )
         )
         // show add tokens button if user is on pro workspace
         .when(
-          () => isProWorkspace,
+          () => isProWorkspace || isTeamsWorkspace,
           () => (
             <Tooltip
               referenceElement={({ ref, onOpen, onClose }) => (
@@ -93,13 +102,22 @@ export const AssistantNavigationTokenUsage: React.FC<BaseProps> = ({ testID }) =
                 />
               )}
             >
-              {() => <Tooltip.Caption mb={0}>Only workspace admins can add more tokens. Ask your admin to purchase a top-up.</Tooltip.Caption>}
+              {() => (
+                <Tooltip.Caption mb={0}>
+                  Only workspace admins can add more tokens. Ask your admin to purchase a top-up.
+                </Tooltip.Caption>
+              )}
             </Tooltip>
           )
         )
         // show contact sales button otherwise
         .otherwise(() => (
-          <Tooltip.Button size="small" label="Contact sales" testID={tid(testID, 'contact-sales')} onClick={onOpenBookDemoPage} />
+          <Tooltip.Button
+            size="small"
+            label="Contact sales"
+            testID={tid(testID, 'contact-sales')}
+            onClick={onOpenBookDemoPage}
+          />
         ))}
     />
   );