+++ libs/orm-designer/src/postgres/reference/reference-resource/reference-resource.orm.ts
@@ -1,0 +1,10 @@
+import { PostgresObjectIDMutableORM } from '@/postgres/common/orms/postgres-object-id-mutable.orm';
+
+import { ReferenceResourceEntity } from './reference-resource.entity';
+import { ReferenceResourceJSONAdapter } from './reference-resource-json.adapter';
+
+export class ReferenceResourceORM extends PostgresObjectIDMutableORM<ReferenceResourceEntity> {
+  Entity = ReferenceResourceEntity;
+
+  jsonAdapter = ReferenceResourceJSONAdapter;
+}