+++ apps/ml-gateway/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.99.8](https://github.com/voiceflow/creator-app/compare/@voiceflow/ml-gateway@1.99.7...@voiceflow/ml-gateway@1.99.8) (2024-06-07)
+
+**Note:** Version bump only for package @voiceflow/ml-gateway
+
+## [1.99.7](https://github.com/voiceflow/creator-app/compare/@voiceflow/ml-gateway@1.99.6...@voiceflow/ml-gateway@1.99.7) (2024-06-07)
+
+### Bug Fixes
+
+* retry integrate org members (COR-2143) ([#8183](https://github.com/voiceflow/creator-app/issues/8183)) ([2b5e0b4](https://github.com/voiceflow/creator-app/commit/2b5e0b4882240a03e64e5f9940cb76d26bf261da))
+
 ## [1.99.6](https://github.com/voiceflow/creator-app/compare/@voiceflow/ml-gateway@1.99.5...@voiceflow/ml-gateway@1.99.6) (2024-06-06)
 
 **Note:** Version bump only for package @voiceflow/ml-gateway