+++ libs/realtime-sdk/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.412.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.411.1...@voiceflow/realtime-sdk@1.412.0) (2024-06-07)
+
+### Features
+
+* load env data via http (COR-000) ([#8230](https://github.com/voiceflow/creator-app/issues/8230)) ([2b8b1a2](https://github.com/voiceflow/creator-app/commit/2b8b1a254bdb9a09f747ff26227d1d79ae7acc66))
+
+## [1.411.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.411.0...@voiceflow/realtime-sdk@1.411.1) (2024-06-07)
+
+### Bug Fixes
+
+* retry integrate org members (COR-2143) ([#8183](https://github.com/voiceflow/creator-app/issues/8183)) ([2b5e0b4](https://github.com/voiceflow/creator-app/commit/2b5e0b4882240a03e64e5f9940cb76d26bf261da))
+
+# [1.411.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.410.0...@voiceflow/realtime-sdk@1.411.0) (2024-06-06)
+
+### Features
+
+* add partnerKey to signup (COR-1891) ([#8152](https://github.com/voiceflow/creator-app/issues/8152)) ([5e6f9de](https://github.com/voiceflow/creator-app/commit/5e6f9de63a45f1b28bafd773e67b8fde07877164))
+
 # [1.410.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.409.0...@voiceflow/realtime-sdk@1.410.0) (2024-06-06)
 
 ### Features