+++ libs/sdk-logux-designer/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.57.19](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.57.18...@voiceflow/sdk-logux-designer@2.57.19) (2024-06-07)
+
+**Note:** Version bump only for package @voiceflow/sdk-logux-designer
+
+## [2.57.18](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.57.17...@voiceflow/sdk-logux-designer@2.57.18) (2024-06-07)
+
+### Bug Fixes
+
+* retry integrate org members (COR-2143) ([#8183](https://github.com/voiceflow/creator-app/issues/8183)) ([2b5e0b4](https://github.com/voiceflow/creator-app/commit/2b5e0b4882240a03e64e5f9940cb76d26bf261da))
+
+## [2.57.17](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.57.16...@voiceflow/sdk-logux-designer@2.57.17) (2024-06-06)
+
+**Note:** Version bump only for package @voiceflow/sdk-logux-designer
+
 ## [2.57.16](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.57.15...@voiceflow/sdk-logux-designer@2.57.16) (2024-06-06)
 
 **Note:** Version bump only for package @voiceflow/sdk-logux-designer