+++ libs/utils-designer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.20.9](https://github.com/voiceflow/creator-app/compare/@voiceflow/utils-designer@1.20.8...@voiceflow/utils-designer@1.20.9) (2024-06-07)
+
+**Note:** Version bump only for package @voiceflow/utils-designer
+
+## [1.20.8](https://github.com/voiceflow/creator-app/compare/@voiceflow/utils-designer@1.20.7...@voiceflow/utils-designer@1.20.8) (2024-06-07)
+
+**Note:** Version bump only for package @voiceflow/utils-designer
+
 ## [1.20.7](https://github.com/voiceflow/creator-app/compare/@voiceflow/utils-designer@1.20.6...@voiceflow/utils-designer@1.20.7) (2024-06-06)
 
 **Note:** Version bump only for package @voiceflow/utils-designer